// frontend/src/features/gloomi-customizer/fabricPatterns/types.ts
/** Claves de estampado procedural (animal print + tartan holiday). */
export type FabricPatternKey =
  | 'animal_leopard_classic'
  | 'animal_tiger_orange'
  | 'animal_cheetah_spots'
  | 'animal_cow_brown'
  | 'animal_cow_black'
  | 'animal_giraffe_light'
  | 'animal_leopard_yellow'
  | 'animal_leopard_dense'
  | 'animal_tiger_tan'
  | 'animal_zebra_fine'
  | 'animal_giraffe_dark'
  | 'animal_zebra_bold'
  | 'tartan_green_white'
  | 'tartan_red_classic'
  | 'tartan_red_gold'
  | 'tartan_forest'
  | 'tartan_soft_green'
  | 'tartan_buffalo'
  | 'tartan_windowpane'
  | 'tartan_red_green_block'
  | 'tartan_green_dense'
  | 'tartan_green_mono'